feat: resolved pr comments
