+++ src/app/components/account/account-balance-label.tsx
@@ -1,0 +1,27 @@
+import { memo } from 'react';
+
+import {
+  AccountBalanceCaption,
+  AccountBalanceLoading,
+} from '@app/components/account/account-balance-caption';
+import { useStxMarketData } from '@app/query/common/market-data/market-data.hooks';
+import { useUnanchoredStacksBalances } from '@app/query/stacks/balance/balance.hooks';
+
+interface AccountBalanceLabelProps {
+  address: string;
+}
+export const AccountBalanceLabel = memo(({ address }: AccountBalanceLabelProps) => {
+  const stxMarketData = useStxMarketData();
+  const { data: balances, isLoading } = useUnanchoredStacksBalances(address);
+
+  if (isLoading) return <AccountBalanceLoading />;
+
+  if (!balances) return null;
+
+  return (
+    <AccountBalanceCaption
+      availableBalance={balances.stx.availableStx}
+      marketData={stxMarketData}
+    />
+  );
+});
