+++ src/app/pages/send-tokens/components/recipient-field/recipient-account-drawer.tsx
@@ -1,12 +1,15 @@
 import { memo } from 'react';
+import { Virtuoso } from 'react-virtuoso';
 
 import { Box } from '@stacks/ui';
 
 import { useWalletType } from '@app/common/use-wallet-type';
-import { ControlledDrawer } from '@app/components/drawer/controlled-drawer';
+import { BaseDrawer } from '@app/components/drawer/base-drawer';
 import { useAccounts } from '@app/store/accounts/account.hooks';
 
 import { RecipientAccountListItem } from './recipient-account-list-item';
+
+const smallNumberOfAccountsToRenderWholeList = 10;
 
 interface RecipientAccountDrawerProps {
   closeAccountsDrawer(): void;
@@ -18,20 +21,34 @@ export const RecipientAccountDrawer = memo(
     const { whenWallet } = useWalletType();
     const accounts = useAccounts();
 
-    return accounts ? (
-      <ControlledDrawer title="My accounts" isShowing={isShowing} onClose={closeAccountsDrawer}>
-        <Box mb={whenWallet({ ledger: 'base', software: '' })} marginBottom={8}>
-          {accounts.map(item => (
-            <Box mx={['base-loose', 'extra-loose']}>
-              <RecipientAccountListItem
-                handleClose={closeAccountsDrawer}
-                account={item}
-                key={item.address}
-              />
+    return accounts && isShowing ? (
+      <BaseDrawer title="My accounts" isShowing={isShowing} onClose={closeAccountsDrawer}>
+        <Box mx={['base-loose', 'extra-loose']}>
+          {accounts.length <= smallNumberOfAccountsToRenderWholeList ? (
+            <Box mb={whenWallet({ ledger: 'base', software: '' })} marginBottom={8}>
+              {accounts.map(item => (
+                <RecipientAccountListItem
+                  handleClose={closeAccountsDrawer}
+                  account={item}
+                  key={item.address}
+                />
+              ))}
             </Box>
-          ))}
+          ) : (
+            <Virtuoso
+              height="72px"
+              style={{ paddingTop: '24px', height: '70vh' }}
+              totalCount={accounts.length}
+              itemContent={index => (
+                <RecipientAccountListItem
+                  handleClose={closeAccountsDrawer}
+                  account={accounts[index]}
+                />
+              )}
+            />
+          )}
         </Box>
-      </ControlledDrawer>
+      </BaseDrawer>
     ) : null;
   }
 );
